requirements.txt
attrs
loguru
PySide6
xdg
setuptools
PyYAML
psutil
chardet